Moved DomainEvents to their appropriate location
