+++ CoreWiki.Application/Articles/Managing/Events/ArticleDeletedNotification.cs
@@ -1,7 +1,7 @@
 ﻿using CoreWiki.Core.Domain;
 using MediatR;
 
-namespace CoreWiki.Application.Articles.Notifications
+namespace CoreWiki.Application.Articles.Managing.Events
 {
 	public class ArticleDeletedNotification : INotification
 	{
