+++ CoreWiki.Application/Articles/Managing/Impl/ArticleManagementService.cs
@@ -3,8 +3,7 @@ using System.Threading.Tasks;
 using AutoMapper;
 using CoreWiki.Application.Articles.Exceptions;
 using CoreWiki.Application.Articles.Managing.Dto;
-using CoreWiki.Application.Articles.Notifications;
-using CoreWiki.Application.Articles.Reading.Dto;
+using CoreWiki.Application.Articles.Managing.Events;
 using CoreWiki.Application.Helpers;
 using CoreWiki.Core.Domain;
 using CoreWiki.Data.Abstractions.Interfaces;
