+++ CoreWiki.Application/Articles/Reading/Impl/ArticleReadingService.cs
@@ -3,8 +3,8 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using AutoMapper;
-using CoreWiki.Application.Articles.Notifications;
 using CoreWiki.Application.Articles.Reading.Dto;
+using CoreWiki.Application.Articles.Reading.Events;
 using CoreWiki.Core.Domain;
 using CoreWiki.Data.Abstractions.Interfaces;
 using MediatR;
