--- CoreWiki.Application/Articles/Validation/ValidationClassesLiveHere.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace CoreWiki.Application.Articles.Validation
-{
-    class ValidationClassesLiveHere
-    {
-    }
-}
